pyproject.toml
[project]
name = "darkwave-taskmgr"
version = "0.1.0"
description = "Natural language task manager powered by Google Gemini and Todoist"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "python-dotenv>=1.1.1",
    "langchain-core>=0.3.72",
    "langchain-google-genai>=2.1.8",
    "langchain>=0.3.27",
    "pydantic>=2.11.7",
    "todoist-api-python>=2.1.0",
    "tzdata>=2024.1",
    "streamlit>=1.29.0",
    "requests>=2.31.0",
    "urllib3>=2.0.0",
    "structlog>=23.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.3",
    "pytest-playwright>=0.7.1",
    "pytest-mock>=3.12.0",
    "responses>=0.24.1",
    "ruff>=0.8.0",
    "mypy>=1.13.0",
    "pre-commit>=4.0.0",
    "types-requests>=2.31.0",
]

[project.scripts]
run-app = "streamlit:run_app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app", "config"]

[tool.ruff]
line-length = 100
target-version = "py313"
exclude = [
    ".git",
    ".venv",
    "__pycache__",
    ".pytest_cache",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "F",   # pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by formatter)
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]  # allow assert in tests
"app/main.py" = ["E402"]  # sys.path manipulation before imports

[tool.ruff.format]
quote-style = "double"
indent-style = "space"

[tool.mypy]
python_version = "3.13"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
strict = false

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
]
markers = [
    "ui: UI tests using Playwright",
    "slow: Slow running tests",
]
